Add edits to task2

+++ src/main/java/edu/hw2/task2/Rectangle.java
@@ -18,6 +18,6 @@ public class Rectangle {
     }
 
     public double area() {
-        return width * height;
+        return (double) width * height;
     }
 }
